requirements.txt
futures==2.1.5
papy==1.0.5
numpy==1.6.1
scipy==0.13.1
psutil==1.2.1
ExifRead==1.4.0
scikit-image==0.10.1
pillow==2.6.1
geographiclib==1.34
astropy==0.4
astroquery==0.2.1
pyephem==3.7.5.2
basemap==1.0.7
matplotlib==1.1.1rc
brewer2mpl==1.3.2
rawpy==0.1.0
lensfunpy==1.0.2